Avoid infinite loop getting items from TicketUtils
Currently, when we retrieve items from ticketutil and there
are no results, the response says the number of TotalPages is 0 and our
current page is 1. Because we stop recursing when TotalPages equals
current page, we enter an infinite loop. This commit fixes that by
checking whether the current page is greater than or equal
to the number of total pages, rather than just for equality.